+++ vrchat_dart/lib/src/model/streaming/streamed_current_user.g.dart
@@ -10,20 +10,19 @@ part of 'streamed_current_user.dart';
 
 StreamedCurrentUser _$StreamedCurrentUserFromJson(Map<String, dynamic> json) =>
     StreamedCurrentUser(
-      bio: json['bio'] as String,
+      bio: json['bio'] as String?,
       currentAvatar: json['currentAvatar'] as String?,
       currentAvatarImageUrl: json['currentAvatarImageUrl'] as String,
       currentAvatarThumbnailImageUrl:
           json['currentAvatarThumbnailImageUrl'] as String,
       displayName: json['displayName'] as String,
-      fallbackAvatar: json['fallbackAvatar'] as String,
+      fallbackAvatar: json['fallbackAvatar'] as String?,
       id: json['id'] as String,
-      profilePicOverride: json['profilePicOverride'] as String,
+      iconUrl: json['iconUrl'] as String?,
       status: $enumDecode(_$UserStatusEnumMap, json['status']),
       statusDescription: json['statusDescription'] as String,
       tags: (json['tags'] as List<dynamic>).map((e) => e as String).toList(),
-      userIcon: json['userIcon'] as String,
-      username: json['username'] as String,
+      username: json['username'] as String?,
     );
 
 Map<String, dynamic> _$StreamedCurrentUserToJson(
@@ -36,11 +35,10 @@ Map<String, dynamic> _$StreamedCurrentUserToJson(
   'displayName': instance.displayName,
   'fallbackAvatar': instance.fallbackAvatar,
   'id': instance.id,
-  'profilePicOverride': instance.profilePicOverride,
+  'iconUrl': instance.iconUrl,
   'status': _$UserStatusEnumMap[instance.status]!,
   'statusDescription': instance.statusDescription,
   'tags': instance.tags,
-  'userIcon': instance.userIcon,
   'username': instance.username,
 };